+++ src/app/slices/settings.ts
@@ -3,9 +3,8 @@ import { createSlice } from "codeforlife/slices"
 
 import { getSettingsCookie } from "../utils"
 
-export const TWO_PANEL_LAYOUTS = ["auto", "vertical", "horizontal"] as const
+export const TWO_PANEL_LAYOUTS = ["vertical", "horizontal"] as const
 export const THREE_PANEL_LAYOUTS = [
-  "auto",
   "verticalWithLeftHorizontal",
   "vertical",
   "horizontal",
@@ -14,17 +13,22 @@ export type TwoPanelLayout = (typeof TWO_PANEL_LAYOUTS)[number]
 export type ThreePanelLayout = (typeof THREE_PANEL_LAYOUTS)[number]
 export type PanelLayout = TwoPanelLayout | ThreePanelLayout
 
+export const PLAY_SPEEDS = [0.5, 0.75, 1, 1.25, 1.5, 1.75, 2] as const
+export type PlaySpeed = (typeof PLAY_SPEEDS)[number]
+
 export interface SettingsState {
-  twoPanelLayout: TwoPanelLayout
-  threePanelLayout: ThreePanelLayout
+  twoPanelLayout?: TwoPanelLayout
+  threePanelLayout?: ThreePanelLayout
+  playSpeed: PlaySpeed
 }
 
 const DEFAULT_SETTINGS: SettingsState = Object.freeze({
-  twoPanelLayout: "auto",
-  threePanelLayout: "auto",
+  twoPanelLayout: undefined,
+  threePanelLayout: undefined,
+  playSpeed: 1,
 })
 
-const settingsSlice = createSlice({
+export const settingsSlice = createSlice({
   name: "settings",
   initialState: getSettingsCookie() ?? DEFAULT_SETTINGS,
   reducers: create => ({
@@ -38,12 +42,13 @@ const settingsSlice = createSlice({
         state.threePanelLayout = action.payload
       },
     ),
+    setPlaySpeed: create.reducer((state, action: PayloadAction<PlaySpeed>) => {
+      state.playSpeed = action.payload
+    }),
   }),
-  selectors: {
-    selectSettings: settings => settings,
-  },
+  selectors: { selectSettings: settings => settings },
 })
 
-export default settingsSlice
-export const { setTwoPanelLayout, setThreePanelLayout } = settingsSlice.actions
+export const { setTwoPanelLayout, setThreePanelLayout, setPlaySpeed } =
+  settingsSlice.actions
 export const { selectSettings } = settingsSlice.selectors