+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.4.0](https://github.com/ocadotechnology/codeforlife-rapid-router-frontend/compare/v1.3.0...v1.4.0) (2026-01-07)
+
+
+### Features
+
+* **level:** add blockly controls ([75a70a4](https://github.com/ocadotechnology/codeforlife-rapid-router-frontend/commit/75a70a4f4054cbbb92cc7707edf10a6f0a9065ae))
+
+# [1.3.0](https://github.com/ocadotechnology/codeforlife-rapid-router-frontend/compare/v1.2.0...v1.3.0) (2025-12-19)
+
+
+### Features
+
+* **Blockly:** add custom blocks ([6f3e022](https://github.com/ocadotechnology/codeforlife-rapid-router-frontend/commit/6f3e0225622addc1dffae327be1e550b4ef9c3b9))
+
 # [1.2.0](https://github.com/ocadotechnology/codeforlife-rapid-router-frontend/compare/v1.1.0...v1.2.0) (2025-11-14)